+++ apps/cowswap-frontend/src/common/hooks/useConfirmPriceImpactWithoutFee.ts
@@ -1,60 +1,55 @@
 import { useCallback, useMemo, useState } from 'react'
 
-import { ALLOWED_PRICE_IMPACT_HIGH, PRICE_IMPACT_WITHOUT_FEE_CONFIRM_MIN } from '@cowprotocol/common-const'
 import { Percent } from '@uniswap/sdk-core'
 
 import { t } from '@lingui/core/macro'
 
 import { useConfirmationRequest } from 'common/hooks/useConfirmationRequest'
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-function getDescription(priceImpactWithoutFee: Percent) {
-  const pct = priceImpactWithoutFee.toFixed(0)
+import { BRIDGE_PRICE_IMPACT_THRESHOLD, PRICE_IMPACT_THRESHOLD } from '../constants/priceImpact'
 
-  if (!priceImpactWithoutFee.lessThan(PRICE_IMPACT_WITHOUT_FEE_CONFIRM_MIN)) {
-    return t`This swap has a price impact of at least ${pct}%.`
-  }
-
-  if (!priceImpactWithoutFee.lessThan(ALLOWED_PRICE_IMPACT_HIGH)) {
-    return t`This swap has a price impact of at least ${pct}%.`
-  }
-
-  return undefined
+interface ConfirmPriceImpactContext {
+  confirmPriceImpactWithoutFee: (priceImpactWithoutFee: Percent | undefined) => Promise<boolean>
+  isConfirmed: boolean
 }
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-function shouldSkipInput(priceImpactWithoutFee: Percent) {
-  return (
-    priceImpactWithoutFee.lessThan(PRICE_IMPACT_WITHOUT_FEE_CONFIRM_MIN) &&
-    !priceImpactWithoutFee.lessThan(ALLOWED_PRICE_IMPACT_HIGH)
-  )
-}
-
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function useConfirmPriceImpactWithoutFee() {
+export function useConfirmPriceImpactWithoutFee(isBridge: boolean): ConfirmPriceImpactContext {
   const [isConfirmed, setIsConfirmed] = useState(false)
   const onEnable = useCallback(() => setIsConfirmed(true), [])
   const triggerConfirmation = useConfirmationRequest({ onEnable })
 
   const confirmPriceImpactWithoutFee = useCallback(
-    async (priceImpactWithoutFee: Percent | undefined) => {
-      if (
-        !!priceImpactWithoutFee &&
-        (!priceImpactWithoutFee.lessThan(PRICE_IMPACT_WITHOUT_FEE_CONFIRM_MIN) ||
-          !priceImpactWithoutFee.lessThan(ALLOWED_PRICE_IMPACT_HIGH))
-      ) {
+    async (priceImpact: Percent | undefined) => {
+      const { high: highThreshold, critical: criticalThreshold } = isBridge
+        ? BRIDGE_PRICE_IMPACT_THRESHOLD
+        : PRICE_IMPACT_THRESHOLD
+
+      if (!!priceImpact && (!priceImpact.lessThan(criticalThreshold) || !priceImpact.lessThan(highThreshold))) {
         setIsConfirmed(false)
         try {
+          const shouldSkipInput = priceImpact.lessThan(criticalThreshold) && !priceImpact.lessThan(highThreshold)
+
+          const description = (() => {
+            const pct = priceImpact.toFixed(0)
+
+            if (!priceImpact.lessThan(criticalThreshold)) {
+              return t`This swap has a price impact of at least ${pct}%.`
+            }
+
+            if (!priceImpact.lessThan(highThreshold)) {
+              return t`This swap has a price impact of at least ${pct}%.`
+            }
+
+            return undefined
+          })()
+
           const result = await triggerConfirmation({
             confirmWord: t`confirm`,
             title: t`Confirm Price Impact`,
             action: t`continue with this swap`,
             callToAction: t`Confirm Swap`,
-            description: getDescription(priceImpactWithoutFee),
-            skipInput: shouldSkipInput(priceImpactWithoutFee),
+            description,
+            skipInput: shouldSkipInput,
           })
           setIsConfirmed(result)
           return result
@@ -67,7 +62,7 @@ export function useConfirmPriceImpactWithoutFee() {
         return true
       }
     },
-    [triggerConfirmation],
+    [triggerConfirmation, isBridge],
   )
 
   return useMemo(