+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/BridgingStatusHeader/index.tsx
@@ -43,11 +43,7 @@ export interface BridgingStatusHeaderProps {
   stepName: BridgingFlowStep
 }
 
-export function BridgingStatusHeader({
-  stepName,
-  sellToken,
-  buyToken,
-}: BridgingStatusHeaderProps): ReactNode {
+export function BridgingStatusHeader({ stepName, sellToken, buyToken }: BridgingStatusHeaderProps): ReactNode {
   const { i18n } = useLingui()
   const isBridgingFailed = stepName === 'bridgingFailed'
   const sellTokenEl = createTokenLogo(sellToken, stepName)