+++ libs/wallet/src/constants.ts
@@ -11,4 +11,6 @@ export const METAMASK_RDNS = 'io.metamask'
 
 export const RABBY_RDNS = 'io.rabby'
 
+export const BRAVE_WALLET_RDNS = 'com.brave.wallet'
+
 export const WATCH_ASSET_SUPPORED_WALLETS = [METAMASK_RDNS]